principle included in all post/put/delete requests; Object ownership checks

+++ src/main/java/com/shelter/mykyda/service/DeleteService.java
@@ -1,5 +1,6 @@
 package com.shelter.mykyda.service;
 
+import com.shelter.mykyda.database.entity.Role;
 import com.shelter.mykyda.database.repository.*;
 import com.shelter.mykyda.exeption.NotFoundException;
 import com.shelter.mykyda.security.repository.UserRepository;
@@ -8,6 +9,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.security.Principal;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -25,37 +29,52 @@ public class DeleteService {
 
     private final MediaService mediaService;
 
-    //:ToDo add Principle and ownership check
     @Transactional(rollbackFor = RuntimeException.class)
-    public ResponseEntity<String> deletePost(Long id) {
-
-        //var user = userRepository.findByEmail(dto.getEmail()).orElseThrow(() -> new NotFoundException("User with such email not found"));
+    public ResponseEntity<String> deletePost(Long id, Principal principal) {
+        var user = userRepository.findByEmail(principal.getName()).orElseThrow(() -> new NotFoundException("User with such email not found"));
         var post = postRepository.findById(id).orElseThrow(() -> new NotFoundException("Post with such id not found"));
-        var images = post.getImages();
-        if (images != null) {
-            images.forEach(image -> {
-                mediaService.deletePostImage(image.getUrl());
-                imageRepository.deleteById(image.getId());
-            });
+        if (Objects.equals(user.getId(), post.getUser().getId())) {
+            var images = post.getImages();
+            if (images != null) {
+                images.forEach(image -> {
+                    mediaService.deletePostImage(image.getUrl());
+                    imageRepository.deleteById(image.getId());
+                });
+            }
+            postRepository.deleteById(id);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
-        postRepository.deleteById(id);
-
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @Transactional(rollbackFor = RuntimeException.class)
-    public ResponseEntity<String> deleteNewsItem(Long id) {
-        //var user = userRepository.findByEmail(dto.getEmail()).orElseThrow(() -> new NotFoundException("User with such email not found"));
-        newsItemRepository.findById(id).orElseThrow(() -> new NotFoundException("News item with such id not found"));
-        newsItemRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<String> deleteNewsItem(Long id, Principal principal) {
+        var user = userRepository.findByEmail(principal.getName()).orElseThrow(() -> new NotFoundException("User with such email not found"));
+        var item = newsItemRepository.findById(id).orElseThrow(() -> new NotFoundException("NewsItem with such id not found"));
+        if (user.getRole() == Role.MANAGER && user.getShelter() == item.getShelter()) {
+            newsItemRepository.findById(id).orElseThrow(() -> new NotFoundException("News item with such id not found"));
+            newsItemRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
     }
 
     @Transactional(rollbackFor = RuntimeException.class)
-    public ResponseEntity<String> deleteShelter(Long id) {
-        //var user = userRepository.findByEmail(dto.getEmail()).orElseThrow(() -> new NotFoundException("User with such email not found"));
-        shelterRepository.findById(id).orElseThrow(() -> new NotFoundException("News item with such id not found"));
-        shelterRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<String> deleteShelter(Long id, Principal principal) {
+        var user = userRepository.findByEmail(principal.getName()).orElseThrow(() -> new NotFoundException("User with such email not found"));
+        if (user.getRole() == Role.MANAGER && user.getShelter()!=null) {
+            if (Objects.equals(user.getShelter().getId(), id)) {
+                shelterRepository.findById(id).orElseThrow(() -> new NotFoundException("News item with such id not found"));
+                shelterRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.CONFLICT);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
     }
 }
